python-backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.88",
    "fastapi[standard]",
    "httpx>=0.28.1",
    "langchain-text-splitters>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "onnxruntime>=1.24.4",
    "openai>=2.31.0",
    "pillow>=12.2.0",
    "planetary-computer>=1.0.0",
    "prisma>=0.15.0",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.2.2",
    "pymupdf4llm>=1.27.2.2",
    "pystac-client>=0.9.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.25",
    "rasterio>=1.5.0",
    "scikit-image>=0.26.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
